Implement picture analyzer in bot
