+++ backend/commands/terms/termsMessage.js
@@ -1,0 +1,37 @@
+const termsMessage = (
+  botName,
+  pointsActions
+) => `<b>Community-Focused Airdrop: GD Token on TON Blockchain</b>
+
+Welcome to GhostDrive Community Airdrop for the GD Token on the Ton Blockchain! Our goal is to reward active users with GD points for their actions and tasks, fostering a vibrant and engaged community.
+
+How It Works:
+
+1. Join the Telegram Program
+   - Connect with our bot: [@${botName}](#).
+   - Engage with the community and stay updated with the latest announcements.
+
+2. Launch GhostDrive Telegram Mini App
+   - Click the "Open GhostDrive" button to launch the app directly within Telegram.
+   - Seamlessly access and manage your GD points and rewards.
+
+3. Start Point Mining
+   - Begin earning GD Points through various activities and tasks.
+   - Track your progress and points accumulation through the bot.
+
+Earning GD Points:
+
+${pointsActions}
+
+Points Booster:
+
+We also offer a Points Booster packages that includes additional storage space for one year. Enhance your earning potential and enjoy expanded storage capabilities:
+
+Points Booster with Storage Space:
+
+  - Unlock increased storage capacity for 12 months.
+  - Boost your GD Points earnings with exclusive benefits.
+
+Ends by Aug 16`;
+
+export default termsMessage;
